refactor(task): split controller validation helpers
